:construction :recycle: :memo: add more structure, basic of swagger, and renaming files.
