+++ api/swagger/swagger_config.py
@@ -1,0 +1,12 @@
+swagger_configuration = {
+    'headers': [],
+    'specs': [
+        {
+            'endpoint': 'swagger',
+            'route': '/swagger.json'
+        }
+    ],
+    'swagger_ui': True,
+    "static_url_path": "/flasgger_static",
+    'specs_route': '/swagger/'
+}
