+++ requirements.txt
@@ -1,4 +1,3 @@
 tweepy=3.7.0
 pymongo=3.8.0
 python-dotenv=0.10.1
-
